😥 update authorization

+++ src/main/java/com/example/froggyblogserver/repository/CommentRepo.java
@@ -20,7 +20,7 @@ public interface CommentRepo extends JpaRepository<CommentEntity, String> {
     @Query(value = "SELECT COUNT(c) FROM CommentEntity c WHERE c.parentId = :id AND c.isDelete = 0")
     Optional<Integer> countByParentId(String id);
     @Query(value = "SELECT COUNT(c) FROM CommentEntity c WHERE c.postId = :postId AND c.isDelete = 0")
-    Optional<Integer> countByPostId(String postId);
+    Optional<Long> countByPostId(String postId);
     @Query(value = "SELECT COUNT(c) FROM CommentEntity c WHERE c.postId = :postId AND c.isDelete = 0 AND c.createDate >= :startDate AND c.createDate <=:endDate" )
     Optional<Long> countByUser(String postId, LocalDateTime startDate,LocalDateTime endDate);
 }
